Basic index.html website project submission
This is my index.html skeleton. I've been writing my code in another index.html file so I can play with it and make mistakes or add random things. I copy pasted a clean version of it for my Push. I deselected many misc. changes as they were pertaining to the addition of audio and image files that may be left unused or deleted later.